src/requirements.txt
flask
numpy==1.19.2
#torch==1.10.2
#torchvision==0.11.2
Pillow==8.0.1
